+++ src/components/index.ts
@@ -56,6 +56,7 @@ export { default as ChatInput } from './molcule/chat-input';
 export { default as ChatScreen } from './molcule/chat-screen';
 export { default as ContainerDonation } from './molcule/container-donation';
 export { default as KebabProductDetail } from './molcule/kebab-product-detail';
+export { default as PasswordInput } from './molcule/password-input';
 //Organism
 export { default as UserSection } from './organism/user-section';
 export { default as ListTradeItems } from './organism/list-trade-items';
@@ -67,3 +68,4 @@ export { default as FormDonationBasic } from './organism/form-donation-basic';
 export { default as FormDonationRequest } from './organism/form-donation-request';
 export { default as CarouselMagazine } from './organism/carousel-magazine';
 export { default as Ranking } from './organism/ranking';
+export { default as LoginForm } from './organism/login-form';
